requirements.txt
Django>=5.2.8
openpyxl>=3.1.2
Pillow>=10.0.0
mistralai>=1.0.0
python-dotenv>=1.0.0
asgiref==3.11.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==5.2.9
et_xmlfile==2.0.0
Flask==3.1.2
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.4
